Rely only on loadable-components to split chunks
Also move zxcvbn import out of `app/` so it's split off.
